refactor: extract duplicate Semaphore class to shared utils
Both edrsr-vectorizer-service and scrape-worker-service had inline
Semaphore implementations. Unified them into utils/semaphore.ts using
the safer "returns release callback" pattern that prevents forgetting
to call release().
