+++ mcp_backend/src/services/edrsr-vectorizer-service.ts
@@ -12,6 +12,7 @@
 import { QdrantClient } from '@qdrant/js-client-rest';
 import { VoyageAIClient, VoyageBatchResult } from '../utils/voyage-client.js';
 import { logger } from '../utils/logger.js';
+import { Semaphore } from '../utils/semaphore.js';
 import { v4 as uuidv4 } from 'uuid';
 
 // ── Constants ────────────────────────────────────────────────────────────────
@@ -95,36 +96,6 @@ function chunkText(text: string, maxChars = MAX_CHUNK_CHARS, overlapWords = CHUN
   }
 
   return chunks;
-}
-
-// ── Semaphore ────────────────────────────────────────────────────────────────
-
-class Semaphore {
-  private current = 0;
-  private queue: Array<() => void> = [];
-
-  constructor(private max: number) {}
-
-  async acquire(): Promise<void> {
-    if (this.current < this.max) {
-      this.current++;
-      return;
-    }
-    return new Promise<void>((resolve) => {
-      this.queue.push(() => {
-        this.current++;
-        resolve();
-      });
-    });
-  }
-
-  release(): void {
-    this.current--;
-    if (this.queue.length > 0) {
-      const next = this.queue.shift()!;
-      next();
-    }
-  }
 }
 
 // ── Service ──────────────────────────────────────────────────────────────────
@@ -270,7 +241,7 @@ export class EdsrVectorizerService {
     for (let i = 0; i < docs.length; i += batchSize) {
       const batch = docs.slice(i, i + batchSize);
       const promise = (async () => {
-        await semaphore.acquire();
+        const release = await semaphore.acquire();
         try {
           const pointIds = await this._vectorizeBatch(batch);
           // Collect results
@@ -278,7 +249,7 @@ export class EdsrVectorizerService {
             result.set(docId, ids);
           });
         } finally {
-          semaphore.release();
+          release();
         }
       })();
       vectorizePromises.push(promise);
@@ -308,11 +279,11 @@ export class EdsrVectorizerService {
     for (let i = 0; i < docs.length; i += batchSize) {
       const batch = docs.slice(i, i + batchSize);
       const promise = (async () => {
-        await semaphore.acquire();
+        const release = await semaphore.acquire();
         try {
           await this._vectorizeBatch(batch);
         } finally {
-          semaphore.release();
+          release();
         }
       })();
       promises.push(promise);
