+++ mcp_backend/src/services/scrape-worker-service.ts
@@ -19,6 +19,7 @@ import type { IEmbeddingPort } from '../domain/ports/index.js';
 import { DocumentService } from '../services/document-service.js';
 import { ZOAdapter } from '../adapters/zo-adapter.js';
 import { logger } from '../utils/logger.js';
+import { Semaphore } from '../utils/semaphore.js';
 import { SectionType } from '../types/index.js';
 
 // --- Types ---
@@ -53,42 +54,6 @@ export interface BulkScrapeJob {
   error_details: string[];
   started_at: string;
   completed_at?: string;
-}
-
-// --- Semaphore ---
-
-class Semaphore {
-  private current = 0;
-  private queue: Array<() => void> = [];
-
-  constructor(private max: number) {}
-
-  async acquire(): Promise<() => void> {
-    if (this.current < this.max) {
-      this.current++;
-      return () => this.release();
-    }
-    return new Promise<() => void>((resolve) => {
-      this.queue.push(() => {
-        this.current++;
-        resolve(() => this.release());
-      });
-    });
-  }
-
-  private release(): void {
-    this.current = Math.max(0, this.current - 1);
-    const next = this.queue.shift();
-    if (next) next();
-  }
-
-  get pending(): number {
-    return this.queue.length;
-  }
-
-  get inFlight(): number {
-    return this.current;
-  }
 }
 
 // --- Service ---
